Fixes bugs introduced in previous commits.
THIS COMMIT BREAKS FRONTEND TESTING,
there is currently no means by which tests can be run. Issue is in adding
`['@babel/preset-env', { targets: { node: 'current' } }]` to the `presets` list in `babel.config.js` in
`b416810`. At the present moment, its unclear why this is a code breaking change. In order to permit
development, this commit, which acts as a reversion, is necessary.

LET THIS COMMIT BE A REMINDER TO ALWAYS RUN CODE BEFORE PUSHING -- for frontend changes, running unit
tests is not sufficient proof that a change isn't code breaking.